requirement.txt
absl-py==2.1.0
contourpy==1.3.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.54.1
fsspec==2024.2.0
grpcio==1.66.2
Jinja2==3.1.3
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==24.1
pillow==10.2.0
protobuf==5.28.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
six==1.16.0
sympy==1.12
tensorboard==2.18.0
tensorboard-data-server==0.7.2
torch==2.4.1+cpu
torchaudio==2.4.1+cpu
torchvision==0.19.1+cpu
typing_extensions==4.9.0
Werkzeug==3.0.4
